requirements.txt
attrs==25.4.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
flasgger==0.9.7.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mistune==3.2.0
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
typing_extensions==4.15.0
Werkzeug==3.1.5
